+++ lib/karafka/embedded.rb
@@ -28,7 +28,7 @@ module Karafka
           Thread.current.name = 'karafka.embedded'
 
           Karafka::Process.tags.add(:execution_mode, 'mode:embedded')
-          Karafka::Server.execution_mode = :embedded
+          Karafka::Server.execution_mode.embedded!
           Karafka::Server.start
         end
       end